+++ src/lsp.rs
@@ -5,8 +5,8 @@ use std::io::{BufRead, BufWriter, Write};
 use std::path::PathBuf;
 
 const COOL_KEYWORDS: &[&str] = &[
-    "def", "class", "struct", "packed", "union", "if", "elif", "else", "while", "for", "in", "not", "and", "or",
-    "return", "break", "continue", "pass", "import", "from", "as", "try", "except", "finally", "raise", "with",
+    "def", "extern", "class", "struct", "packed", "union", "if", "elif", "else", "while", "for", "in", "not", "and",
+    "or", "return", "break", "continue", "pass", "import", "from", "as", "try", "except", "finally", "raise", "with",
     "lambda", "assert", "global", "nonlocal", "True", "False", "None",
 ];